+++ homeassistant/components/light/lifx.py
@@ -96,8 +96,8 @@ LIFX_EFFECT_PULSE_SCHEMA = LIFX_EFFECT_SCHEMA.extend({
                 vol.Coerce(tuple)),
     vol.Exclusive(ATTR_HS_COLOR, COLOR_GROUP):
         vol.All(vol.ExactSequence(
-            (vol.All(vol.Coerce(int), vol.Range(min=0, max=360)),
-             vol.All(vol.Coerce(int), vol.Range(min=0, max=100)))),
+            (vol.All(vol.Coerce(float), vol.Range(min=0, max=360)),
+             vol.All(vol.Coerce(float), vol.Range(min=0, max=100)))),
                 vol.Coerce(tuple)),
     vol.Exclusive(ATTR_COLOR_TEMP, COLOR_GROUP):
         vol.All(vol.Coerce(int), vol.Range(min=1)),
